Moves StatelessWorkflow, Workflows.kt methods to Workflow.
Hangs methods off extensions of Workflow.companion.
Deletes Workflows.kt.
