+++ kotlin/workflow-testing/src/test/java/com/squareup/workflow/WorkflowTesterTest.kt
@@ -27,7 +27,7 @@ import kotlin.test.fail
 class WorkflowTesterTest {
 
   @Test fun `propagates exception when block throws`() {
-    val workflow = StatelessWorkflow<Unit, Unit> { }
+    val workflow = Workflow.stateless<Unit, Unit> { }
 
     assertFailsWith<ExpectedException> {
       workflow.testFromStart {
@@ -37,7 +37,7 @@ class WorkflowTesterTest {
   }
 
   @Test fun `propagates exception when workflow throws from compose`() {
-    val workflow = StatelessWorkflow<Unit, Unit> {
+    val workflow = Workflow.stateless<Unit, Unit> {
       throw ExpectedException()
     }
 
@@ -50,7 +50,7 @@ class WorkflowTesterTest {
 
   @Test fun `propagates exception when Job is cancelled in test block`() {
     val job = Job()
-    val workflow = StatelessWorkflow<Unit, Unit> { }
+    val workflow = Workflow.stateless<Unit, Unit> { }
 
     assertFailsWith<ClosedReceiveChannelException> {
       workflow.testFromStart(context = job) {
@@ -62,7 +62,7 @@ class WorkflowTesterTest {
 
   @Test fun `propagates exception when Job is cancelled before starting`() {
     val job = Job().apply { cancel() }
-    val workflow = StatelessWorkflow<Unit, Unit> { }
+    val workflow = Workflow.stateless<Unit, Unit> { }
 
     assertFailsWith<JobCancellationException> {
       workflow.testFromStart(context = job) {
@@ -164,7 +164,7 @@ class WorkflowTesterTest {
   @Test fun `propagates exception when subscription throws`() {
     val deferred = CompletableDeferred<Unit>()
     deferred.completeExceptionally(ExpectedException())
-    val workflow = StatelessWorkflow<Unit, Unit> {
+    val workflow = Workflow.stateless<Unit, Unit> {
       it.onDeferred(deferred) { fail("Shouldn't get here.") }
     }
 
@@ -176,7 +176,7 @@ class WorkflowTesterTest {
   }
 
   @Test fun `does nothing when no outputs observed`() {
-    val workflow = StatelessWorkflow<Unit, Unit> {}
+    val workflow = Workflow.stateless<Unit, Unit> {}
 
     workflow.testFromStart {
       // The workflow should never start.
